+++ apps/api/src/controllers/v2/types.ts
@@ -686,7 +686,10 @@ const baseScrapeOptions = z.strictObject({
   fastMode: z.boolean().prefault(false),
   useMock: z.string().optional(),
   blockAds: z.boolean().prefault(true),
-  proxy: z.enum(["basic", "stealth", "enhanced", "auto"]).prefault("auto"),
+  proxy: z
+    .enum(["basic", "stealth", "enhanced", "auto"])
+    .prefault("auto")
+    .transform(v => (v === "stealth" ? "enhanced" : v)),
   maxAge: z.int().gte(0).optional(),
   minAge: z.int().gte(0).optional(),
   storeInCache: z.boolean().prefault(true),
@@ -764,9 +767,7 @@ const extractTransformImpl = <T extends ScrapeOptionsBase | undefined>(
   }
 
   if (
-    (obj.proxy === "stealth" ||
-      obj.proxy === "enhanced" ||
-      obj.proxy === "auto") &&
+    (obj.proxy === "enhanced" || obj.proxy === "auto") &&
     obj.timeout === 30000
   ) {
     result = { ...result, timeout: 120000 };
