+++ apps/api/src/lib/error.ts
@@ -6,16 +6,13 @@ export type ErrorCodes =
   | "SCRAPE_SSL_ERROR"
   | "SCRAPE_SITE_ERROR"
   | "SCRAPE_PROXY_SELECTION_ERROR"
-  | "SCRAPE_PDF_PREFETCH_FAILED"
-  | "SCRAPE_DOCUMENT_PREFETCH_FAILED"
   | "SCRAPE_JOB_CANCELLED"
   | "SCRAPE_RETRY_LIMIT"
   | "SCRAPE_ZDR_VIOLATION_ERROR"
   | "SCRAPE_DNS_RESOLUTION_ERROR"
   | "SCRAPE_PDF_INSUFFICIENT_TIME_ERROR"
-  | "SCRAPE_PDF_ANTIBOT_ERROR"
   | "SCRAPE_PDF_OCR_REQUIRED"
-  | "SCRAPE_DOCUMENT_ANTIBOT_ERROR"
+  | "SCRAPE_RELIABLE_RETRIEVAL_ERROR"
   | "SCRAPE_UNSUPPORTED_FILE_ERROR"
   | "SCRAPE_ACTION_ERROR"
   | "SCRAPE_RACED_REDIRECT_ERROR"
