+++ apps/api/src/lib/keyless-credit-projection.ts
@@ -54,11 +54,7 @@ export function projectScrapeCredits(
     credits += 4;
   }
 
-  if (
-    options.proxy === "stealth" ||
-    options.proxy === "enhanced" ||
-    options.proxy === "auto"
-  ) {
+  if (options.proxy === "enhanced" || options.proxy === "auto") {
     credits += 4;
   }
