+++ apps/api/src/scraper/scrapeURL/engines/fire-engine/delete.ts
@@ -1,13 +1,11 @@
 import { Logger } from "winston";
 
 import { robustFetch } from "../../lib/fetch";
-import { MockState } from "../../lib/mock";
 import { fireEngineURL } from "./scrape";
 
 export async function fireEngineDelete(
   logger: Logger,
   jobId: string | undefined,
-  mock: MockState | null,
   abort?: AbortSignal,
   baseUrl: string = fireEngineURL,
 ) {
@@ -22,7 +20,6 @@ export async function fireEngineDelete(
     method: "DELETE",
     headers: {},
     logger: logger.child({ method: "fireEngineDelete/robustFetch", jobId }),
-    mock,
     abort,
   });
