+++ apps/api/src/scraper/scrapeURL/engines/x-twitter/index.ts
@@ -1,10 +1,10 @@
 import { xai } from "@ai-sdk/xai";
 import { generateText, jsonSchema, Output } from "ai";
 import { config } from "../../../../config";
-import { Meta } from "../..";
-import { EngineScrapeResult } from "..";
 import { EngineError, XTwitterConfigurationError } from "../../error";
 import { safeMarkdownToHtml } from "../pdf/markdownToHtml";
+import { Meta } from "../../lib/meta";
+import { Engine, EngineScrapeResult, SpecialEngine } from "../types";
 
 const XAI_RESPONSES_MODEL = "grok-4-1-fast-non-reasoning";
 
@@ -343,10 +343,6 @@ export async function scrapeURLWithXTwitter(
   };
 }
 
-export function xTwitterMaxReasonableTime(_meta: Meta): number {
-  return 30000;
-}
-
 function buildProfileMarkdown(profile: XTwitterProfileData): string {
   const username = stripAt(profile.username) ?? "unknown";
   const displayName = profile.displayName?.trim() || `@${username}`;
@@ -654,3 +650,34 @@ function escapeMarkdownBlock(value: string): string {
 function trimMarkdown(markdown: string): string {
   return markdown.replace(/\n{3,}/g, "\n\n").trim() + "\n";
 }
+
+export const xTwitterSpecialEngine: SpecialEngine = {
+  name: "x-twitter",
+  features: {
+    actions: false,
+    waitFor: false,
+    screenshot: false,
+    "screenshot@fullScreen": false,
+    audio: false,
+    video: false,
+    atsv: false,
+    location: false,
+    mobile: false,
+    branding: false,
+    disableAdblock: true,
+  },
+  scrape: meta => {
+    const logger = meta.logger.child({
+      function: "scrapeURLWithXTwitter",
+      engine: "x-twitter",
+    });
+
+    return scrapeURLWithXTwitter({
+      ...meta,
+      logger,
+    });
+  },
+  special: {
+    regex: /hi/, // TODO:
+  },
+};
