+++ apps/api/src/scraper/scrapeURL/lib/fetch.ts
@@ -1,8 +1,6 @@
 import { Logger } from "winston";
 import { z, ZodError } from "zod";
 import * as Sentry from "@sentry/node";
-import { MockState, saveMock } from "./mock";
-import { fireEngineURL } from "../engines/fire-engine/scrape";
 import { fetch, Response, FormData, Agent } from "undici";
 import { cacheableLookup } from "./cacheableLookup";
 import dns from "dns";
@@ -22,7 +20,6 @@ type RobustFetchParams<Schema extends z.Schema<any>> = {
   requestId?: string;
   tryCount?: number;
   tryCooldown?: number;
-  mock: MockState | null;
   abort?: AbortSignal;
   useCacheableLookup?: boolean;
 };
@@ -59,7 +56,6 @@ export async function robustFetch<
   requestId = crypto.randomUUID(),
   tryCount = 1,
   tryCooldown,
-  mock,
   abort,
   useCacheableLookup = true,
 }: RobustFetchParams<Schema>): Promise<Output> {
@@ -106,113 +102,76 @@ export async function robustFetch<
     body: string;
   };
 
-  if (mock === null) {
-    let request: Response;
-    try {
-      request = await fetch(url, {
-        method,
-        headers: {
-          ...(body instanceof FormData
-            ? {}
-            : body !== undefined
-              ? {
-                  "Content-Type": "application/json",
-                }
-              : {}),
-          ...(headers !== undefined ? headers : {}),
-        },
-        signal: abort,
-        dispatcher: useCacheableLookup ? robustAgent : robustAgentNoLookup,
+  let request: Response;
+  try {
+    request = await fetch(url, {
+      method,
+      headers: {
         ...(body instanceof FormData
-          ? {
-              body,
-            }
+          ? {}
           : body !== undefined
             ? {
-                body: JSON.stringify(body),
+                "Content-Type": "application/json",
               }
             : {}),
-      });
-    } catch (error) {
-      if (error instanceof AbortManagerThrownError) {
-        throw error;
-      } else if (!ignoreFailure) {
-        Sentry.captureException(error);
-        if (tryCount > 1) {
-          logger.debug(
-            "Request failed, trying " + (tryCount - 1) + " more times",
-            { params: logParams, error, requestId },
-          );
-          return await robustFetch({
-            ...params,
-            requestId,
-            tryCount: tryCount - 1,
-            mock,
-          });
-        } else {
-          logger.debug("Request failed", {
-            params: logParams,
-            error,
-            requestId,
-          });
-          throw new Error("Request failed", {
-            cause: {
-              params: logParams,
-              requestId,
-              error,
-            },
-          });
-        }
+        ...(headers !== undefined ? headers : {}),
+      },
+      signal: abort,
+      dispatcher: useCacheableLookup ? robustAgent : robustAgentNoLookup,
+      ...(body instanceof FormData
+        ? {
+            body,
+          }
+        : body !== undefined
+          ? {
+              body: JSON.stringify(body),
+            }
+          : {}),
+    });
+  } catch (error) {
+    if (error instanceof AbortManagerThrownError) {
+      throw error;
+    } else if (!ignoreFailure) {
+      Sentry.captureException(error);
+      if (tryCount > 1) {
+        logger.debug(
+          "Request failed, trying " + (tryCount - 1) + " more times",
+          { params: logParams, error, requestId },
+        );
+        return await robustFetch({
+          ...params,
+          requestId,
+          tryCount: tryCount - 1,
+        });
       } else {
-        return null as Output;
+        logger.debug("Request failed", {
+          params: logParams,
+          error,
+          requestId,
+        });
+        throw new Error("Request failed", {
+          cause: {
+            params: logParams,
+            requestId,
+            error,
+          },
+        });
       }
-    }
-
-    if (ignoreResponse === true) {
+    } else {
       return null as Output;
     }
-
-    const resp = await request.text();
-    response = {
-      status: request.status,
-      headers: request.headers,
-      body: resp, // NOTE: can this throw an exception?
-    };
-  } else {
-    if (ignoreResponse === true) {
-      return null as Output;
-    }
-
-    const makeRequestTypeId = (
-      request: (typeof mock)["requests"][number]["options"],
-    ) => {
-      let trueUrl = request.url.startsWith(fireEngineURL)
-        ? request.url.replace(fireEngineURL, "<fire-engine>")
-        : request.url;
-
-      let out = trueUrl + ";" + request.method;
-      if (trueUrl.startsWith("<fire-engine>") && request.method === "POST") {
-        out += "f-e;" + request.body?.engine + ";" + request.body?.url;
-      }
-      return out;
-    };
-
-    const thisId = makeRequestTypeId(params);
-    const matchingMocks = mock.requests
-      .filter(x => makeRequestTypeId(x.options) === thisId)
-      .sort((a, b) => a.time - b.time);
-    const nextI = mock.tracker[thisId] ?? 0;
-    mock.tracker[thisId] = nextI + 1;
-
-    if (!matchingMocks[nextI]) {
-      throw new Error("Failed to mock request -- no mock targets found.");
-    }
-
-    response = {
-      ...matchingMocks[nextI].result,
-      headers: new Headers(matchingMocks[nextI].result.headers),
-    };
   }
+
+  if (ignoreResponse === true) {
+    return null as Output;
+  }
+
+  const resp = await request.text();
+  response = {
+    status: request.status,
+    headers: request.headers,
+    body: resp, // NOTE: can this throw an exception?
+  };
 
   if (response.status >= 300 && !ignoreFailureStatus) {
     if (tryCount > 1) {
@@ -240,7 +199,6 @@ export async function robustFetch<
         ...params,
         requestId,
         tryCount: tryCount - 1,
-        mock,
       });
     } else {
       logger.debug("Request sent failure status", {
@@ -256,18 +214,6 @@ export async function robustFetch<
         },
       });
     }
-  }
-
-  if (mock === null) {
-    await saveMock(
-      {
-        ...params,
-        logger: undefined,
-        schema: undefined,
-        headers: undefined,
-      },
-      response,
-    );
   }
 
   let data: Output;
