+++ apps/api/src/services/monitoring/store.ts
@@ -128,10 +128,7 @@ function estimateBaseCreditsPerPage(
     credits += SCRAPE_OPTION_CREDIT_BONUS;
   }
 
-  if (
-    includeProxy &&
-    (options?.proxy === "stealth" || options?.proxy === "enhanced")
-  ) {
+  if (includeProxy && options?.proxy === "enhanced") {
     credits += SCRAPE_OPTION_CREDIT_BONUS;
   }
